remove use of as.single
git-svn-id: https://svn.r-project.org/R/trunk@60017 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/stats/R/shapiro.test.R
@@ -30,7 +30,7 @@ shapiro.test <- function(x)
     n2 <- n %/% 2L # integer, too
     ## C Code: Use the first n1 observations as uncensored
     sw <- .C(C_swilk,
-	     as.single(x),
+	     as.double(x),
 	     n,#      integer
 	     n1 = n,#   "
 	     w	= double(1),
